Fix for some Visual C++ x64 compile warnings

+++ tests/DCPS/LargeSample/DataReaderListener.cpp
@@ -217,12 +217,12 @@ bool DataReaderListenerImpl::data_consistent() const
         if (writer->second.size() < samples_per_writer_) {
           for (size_t sample = 0; sample < samples_per_writer_; ++sample) {
             std::cerr
-              << (writer->second.count(sample) ? "           Got" : "ERROR: Missing")
+              << (writer->second.count(static_cast<int>(sample)) ? "           Got" : "ERROR: Missing")
               << " process " << process->first
               << " writer " << writer->first
               << " sample " << sample
               << " expected data length " << expected_data_field_length(
-                data_field_length_offset_, sample, writer->first)
+                data_field_length_offset_, static_cast<int>(sample), writer->first)
               << std::endl;
           }
         }
